Explicit casts.

+++ spine-c/src/spine/Json.c
@@ -74,12 +74,12 @@ static const char* parse_number (Json *item, const char* num) {
 	if (*num == '-') sign = -1, num++; /* Has sign? */
 	if (*num == '0') num++; /* is zero */
 	if (*num >= '1' && *num <= '9') do
-		n = (n * 10.0) + (*num++ - '0');
+		n = (n * 10.0f) + (*num++ - '0');
 	while (*num >= '0' && *num <= '9'); /* Number? */
 	if (*num == '.' && num[1] >= '0' && num[1] <= '9') {
 		num++;
 		do
-			n = (n * 10.0) + (*num++ - '0'), scale--;
+			n = (n * 10.0f) + (*num++ - '0'), scale--;
 		while (*num >= '0' && *num <= '9');
 	} /* Fractional part? */
 	if (*num == 'e' || *num == 'E') /* Exponent? */
@@ -371,7 +371,7 @@ float Json_getFloat (Json* value, const char* name, float defaultValue) {
 
 int Json_getInt (Json* value, const char* name, int defaultValue) {
 	value = Json_getItem(value, name);
-	return value ? value->valuefloat : defaultValue;
+	return value ? (int)value->valuefloat : defaultValue;
 }
 
 #ifdef __cplusplus
